🪓  drop to one demo file, no `redux` required
