--- demo/demo-basic-local.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <link rel="stylesheet" href="https://unpkg.com/@curvenote/article/dist/curvenote.css" />
-    <link rel="stylesheet" href="demo.css" />
-    <script type="text/x-mathjax-config" src="mathjax.js"></script>
-    <script
-      type="text/javascript"
-      async
-      src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/latest.js?config=TeX-MML-AM_CHTML"
-    ></script>
-    <script
-      type="text/javascript"
-      async
-      src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.6/require.min.js"
-    />
-    <script src="https://unpkg.com/@curvenote/article"></script>
-    <script src="demo-basic.js" type="text/javascript"></script>
-  </head>
-  <body>
-    <h1>Demo: thebe-core</h1>
-    <p>
-      This is a demonstration of <em>thebe-core</em> executing an sequence of code cells via the
-      Notebook interface
-    </p>
-    <p>
-      You'll need to start your local jupyter server on port 8888 and like this:
-      <code>jupyter notebook --NotebookApp.token=test-secret --NotebookApp.allow_origin="*"</code>
-    </p>
-    <p>
-      The code itself is in 3 cells as follows, but here it is actually injectes via javascript,
-      rather than picked up from the page.
-    </p>
-    <pre>
-%matplotlib inline
-import numpy as np
-import matplotlib.pyplot as plt
-import scipy.stats as stats</pre
-    >
-    <pre>
-# oxa:fs
-fs = 1
-# oxa:phi
-phi = np.pi/5</pre
-    >
-    <pre>
-plt.ion()
-fig, ax = plt.subplots()
-x = np.arange(-np.pi/2,np.pi/2,0.01)
-y = np.cos(2*np.pi*fs*x+phi)
-yy = np.sin(2*np.pi*fs*x+phi)
-ax.plot(x,y);
-ax.plot(x,yy);
-plt.grid(True)</pre
-    >
-    <p>
-      The page immediately requests a kernel and the kernel status can be monitors via a redux
-      subscription.
-    </p>
-    <p>
-      Current server status is
-      <span id="server-status" class="status">none</span> and session/kernel status is
-      <span id="session-status" class="status">none</span>
-    </p>
-    <button id="connect" style="margin-bottom: 4px; display: block">
-      Connect to Local Jupyter
-    </button>
-    <button id="run-all" disabled>run all</button>
-    <button id="run-last" disabled>run last</button>
-    <button id="restart" disabled>restart</button>
-    <p />
-    <div data-output="true">
-      <img src="/output.png" />
-    </div>
-    <button id="reset" title="removes servers/kernels and force binder to rebuild" disabled>
-      hard reset
-    </button>
-    <script>
-      const options = {
-        useBinder: false,
-        requestSession: true,
-        kernelOptions: {
-          name: 'python3',
-          serverSettings: {
-            appendToken: true,
-            baseUrl: 'http://localhost:8888',
-            token: 'test-secret',
-          },
-        },
-      };
-
-      document.addEventListener('DOMContentLoaded', async function () {
-        const attachButton = document.getElementById('connect');
-        attachButton.onclick = (ev) => {
-          attachButton.setAttribute('disabled', true);
-          demoBasic(code, options);
-        };
-      });
-    </script>
-  </body>
-</html>
