+++ demo/index.html
@@ -16,31 +16,10 @@
       <p>This is a listing of the available demos for thebe-core in use from the native js api,</p>
       <ul>
         <li>
-          A basic example with a matplotlib plot
-          <ul>
-            <li><a href="./basic-binder.html">Connecting to mybinder.org</a></li>
-            <li><a href="./basic-local.html">Connecting to a local Jupyter server</a></li>
-            <li><a href="./basic-lite.html">Connecting to a JupyterLite kernel</a></li>
-          </ul>
-        </li>
-        <li>
-          <a href="./ipywidgets.html">ipywidgets demo connecting to mybinder.org</a>
-        </li>
-        <li>
-          <a href="./local.html">ipywidgets demo connecting to a local Jupyter instance</a>
+          <a href="./demo.html">The main interactive demo</a>
         </li>
         <li>
           <a href="./ipyleaflet.html">ipyleaflet demo connecting to a local Jupyter instance</a>
-        </li>
-        <li>
-          <a href="./demo-single.html"
-            >Demo of rendering a single cell without use of the Notebook interface</a
-          >
-        </li>
-        <li>
-          <a href="./demo-notebook.html"
-            >Demo of rendering a single cell without use of the Notebook interface</a
-          >
         </li>
         <li>
           <a href="./demo-kernels.html"
