--- demo/local.html
@@ -1,141 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <script type="text/x-mathjax-config">
-      MathJax.Hub.Config({
-        extensions: ["tex2jax.js"],
-        jax: ["input/TeX", "output/HTML-CSS"],
-        tex2jax: {
-          inlineMath: [ ['$','$'], ["\\(","\\)"] ],
-          displayMath: [ ['$$','$$'], ["\\[","\\]"] ],
-          processEscapes: true
-        },
-        "HTML-CSS": { fonts: ["TeX"] }
-      });
-    </script>
-    <script
-      type="text/javascript"
-      async
-      src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/latest.js?config=TeX-MML-AM_CHTML"
-    ></script>
-    <script
-      type="text/javascript"
-      async
-      src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.6/require.min.js"
-    />
-    <link
-      rel="stylesheet"
-      href="https://unpkg.com/@curvenote/article/dist/curvenote.css"
-    />
-    <script src="https://unpkg.com/@curvenote/article"></script>
-    <style>
-      body {
-        padding: 10px;
-      }
-      .console {
-        max-height: 300px;
-        position: absolute;
-        bottom: 0;
-        left: 0;
-        border: 2px solid grey;
-        background-color: black;
-        color: greenyellow;
-        width: 100%;
-        max-height: 200px;
-        overflow-y: auto;
-      }
-    </style>
-  </head>
-  <body>
-    <h1>Demo: Local Server: thebe-core with ipywidgets</h1>
-    <p>
-      Ths is a demostration of <em>thebe-core</em> executing an isolated code
-      cell
-    </p>
-    <button id="run-all">run all</button>
-    <button id="run-last">run last</button>
-    <button id="restart">restart</button>
-    <p />
-    <div data-output="true">
-      <img src="/output.png" />
-    </div>
-    <button
-      id="reset"
-      title="removes servers/kernels and force binder to rebuild"
-    >
-      hard reset
-    </button>
-    <script src="/demo.js"></script>
-    <script>
-      const options = {
-        useBinder: false,
-        requestKernel: true,
-        kernelOptions: {
-          name: "python3",
-          serverSettings: {
-            appendToken: true,
-            baseUrl: "http://localhost:8888",
-            token: "test-secret",
-          },
-        },
-      };
-
-      const code = [
-        `%matplotlib widget
-import ipywidgets as widgets
-import matplotlib.pyplot as plt
-import numpy as np
-
-x = np.linspace(0,10)
-
-def sine_func(x, w, amp):
-  return amp*np.sin(w*x)
-
-@widgets.interact(w=(0, 4, 0.25), amp=(0, 4, .1))
-def update(w = 1, amp = 1):
-  plt.clf()
-  plt.ylim(-4, 4)
-  plt.plot(x, sine_func(x, w, amp))`,
-      ];
-
-      document.addEventListener("DOMContentLoaded", function () {
-        function randomId() {
-          const uint32 = window.crypto.getRandomValues(new Uint32Array(1))[0];
-          return uint32.toString(16);
-        }
-
-        window.demo = { notebook: null, kernelId: null };
-        bootstrap(
-          window.demo,
-          options,
-          code.map((c) => ({
-            id: randomId(),
-            source: c,
-          }))
-        );
-
-        const runAllButton = document.getElementById("run-all");
-        runAllButton.onclick = (ev) => {
-          window.demo.notebook.executeAll();
-        };
-
-        const runLastButton = document.getElementById("run-last");
-        runLastButton.onclick = (ev) => {
-          window.demo.notebook.executeOnly(window.demo.notebook.lastCell().id);
-        };
-
-        const restartButton = document.getElementById("restart");
-        restartButton.onclick = (ev) => {
-          window.thebeCore.api.restartKernel(demo.kernelId);
-        };
-
-        const resetButton = document.getElementById("reset");
-        resetButton.onclick = (ev) => {
-          window.thebeCore.api.clear();
-          window.location.reload();
-        };
-      });
-    </script>
-  </body>
-</html>
